requirements.txt
ultralytics>=8.3.50
torch>=2.3.0
opencv-python>=4.8.0.76
numpy>=1.24.0
matplotlib>=3.7.1
seaborn>=0.12.2
pandas>=2.0.0
scikit-learn>=1.3.0
albumentations>=1.4.1
Pillow>=9.5.0
paddlepaddle>=2.6.1
paddleocr>=2.7.1
easyocr>=1.7.1
Flask>=3.0.0
Jinja2>=3.1.2
PyYAML>=6.0.1
flask-cors>=4.0.0
flask-socketio>=5.3.5
python-socketio>=5.10.0
eventlet>=0.33.3
redis>=5.0.0
hiredis>=2.2.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-redis>=2.2.0
requests>=2.31.0
Flask-SQLAlchemy>=3.0.5
python-telegram-bot>=20.0
python-dotenv>=1.0.0



